--- src/main/java/com/sparta/oishitable/global/security/enums/TokenType.java
@@ -1,5 +1,0 @@
-package com.sparta.oishitable.global.security.enums;
-
-public enum TokenType {
-    ACCESS, REFRESH
-}
