+++ src/main/java/com/sparta/oishitable/global/security/model/TokenType.java
@@ -1,0 +1,5 @@
+package com.sparta.oishitable.global.security.model;
+
+public enum TokenType {
+    ACCESS, REFRESH
+}
